requirements.txt
flask
flask-cors
beautifulsoup4
requests
openai
google-generativeai
python-dotenv
flask_sqlalchemy
playwright
firebase-admin
schedule

